+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Support for (P)
 * fix to remove multiple spaces, even they are different
 * fix a.m. " error
+* improve identification of single quotes 
 
 ## 1.4.0 // 2017-01-15
 * “Removing empty lines” is now an optional parameter (in case you want to correct your plain-text based content or an email). By default, empty lines are removed